mint_auto/requirements.txt
numpy==1.18.1
peerplays==0.3.12
PyYAML==5.3
